+++ src/core/soiltemperature.cpp
@@ -272,20 +272,20 @@ void SoilTemperature::serialize(mas::models::monica::SoilTemperatureModuleState:
 	_soilColumn_vt_GroundLayer.serialize(builder.initSoilColumnVtGroundLayer());
 	_soilColumn_vt_BottomLayer.serialize(builder.initSoilColumnVtBottomLayer());
 	_params.serialize(builder.initModuleParams());
-	builder.setNumberOfLayers(vt_NumberOfLayers);
-	builder.setVsNumberOfLayers(vs_NumberOfLayers);
-	setCapnpList(vs_SoilMoisture_const, builder.initVsSoilMoistureConst(vs_SoilMoisture_const.size()));
-	setCapnpList(vt_SoilTemperature, builder.initSoilTemperature(vt_SoilTemperature.size()));
-	setCapnpList(vt_V, builder.initV(vt_V.size()));
-	setCapnpList(vt_VolumeMatrix, builder.initVolumeMatrix(vt_VolumeMatrix.size()));
-	setCapnpList(vt_VolumeMatrixOld, builder.initVolumeMatrixOld(vt_VolumeMatrixOld.size()));
-	setCapnpList(vt_B, builder.initB(vt_B.size()));
-	setCapnpList(vt_MatrixPrimaryDiagonal, builder.initMatrixPrimaryDiagonal(vt_MatrixPrimaryDiagonal.size()));
-	setCapnpList(vt_MatrixSecundaryDiagonal, builder.initMatrixSecundaryDiagonal(vt_MatrixSecundaryDiagonal.size()));
+	builder.setNumberOfLayers((uint16_t)vt_NumberOfLayers);
+	builder.setVsNumberOfLayers((uint16_t)vs_NumberOfLayers);
+	setCapnpList(vs_SoilMoisture_const, builder.initVsSoilMoistureConst((uint)vs_SoilMoisture_const.size()));
+	setCapnpList(vt_SoilTemperature, builder.initSoilTemperature((uint)vt_SoilTemperature.size()));
+	setCapnpList(vt_V, builder.initV((uint)vt_V.size()));
+	setCapnpList(vt_VolumeMatrix, builder.initVolumeMatrix((uint)vt_VolumeMatrix.size()));
+	setCapnpList(vt_VolumeMatrixOld, builder.initVolumeMatrixOld((uint)vt_VolumeMatrixOld.size()));
+	setCapnpList(vt_B, builder.initB((uint)vt_B.size()));
+	setCapnpList(vt_MatrixPrimaryDiagonal, builder.initMatrixPrimaryDiagonal((uint)vt_MatrixPrimaryDiagonal.size()));
+	setCapnpList(vt_MatrixSecundaryDiagonal, builder.initMatrixSecundaryDiagonal((uint)vt_MatrixSecundaryDiagonal.size()));
 	builder.setHeatFlow(vt_HeatFlow);
-	setCapnpList(vt_HeatConductivity, builder.initHeatConductivity(vt_HeatConductivity.size()));
-	setCapnpList(vt_HeatConductivityMean, builder.initHeatConductivityMean(vt_HeatConductivityMean.size()));
-	setCapnpList(vt_HeatCapacity, builder.initHeatCapacity(vt_HeatCapacity.size()));
+	setCapnpList(vt_HeatConductivity, builder.initHeatConductivity((uint)vt_HeatConductivity.size()));
+	setCapnpList(vt_HeatConductivityMean, builder.initHeatConductivityMean((uint)vt_HeatConductivityMean.size()));
+	setCapnpList(vt_HeatCapacity, builder.initHeatCapacity((uint)vt_HeatCapacity.size()));
 	builder.setDampingFactor(_dampingFactor);
 }
